# Rewrite callbacks to dry events