--- a/.travis.yml
+++ b/.travis.yml
@@ -10,6 +10,7 @@ rvm:
   - 2.4.1
   - 2.4.2
   - 2.5.0
+  - 2.5.1
   - jruby-head
 script: bundle exec rspec spec/
 env: