--- a/.travis.yml
+++ b/.travis.yml
@@ -1,16 +1,10 @@
 language: ruby
 sudo: false
 rvm:
-  - 2.3.0
-  - 2.3.1
-  - 2.3.2
-  - 2.3.3
-  - 2.3.4
-  - 2.4.0
-  - 2.4.1
-  - 2.4.2
-  - 2.5.0
+  - 2.3.7
+  - 2.4.3
   - 2.5.1
+  - ruby-head
   - jruby-head
 script: bundle exec rspec spec/
 env: