--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Karafka framework changelog
 
-## Unreleased
+## 1.2.1
 - #304 - Unification of error instrumentation event details
 - #306 - Using file logger from within a trap context upon shutdown is impossible