--- a/README.md
+++ b/README.md
@@ -8,8 +8,6 @@ Karafka allows you to capture everything that happens in your systems in large s
 
 Karafka not only handles incoming messages but also provides tools for building complex data-flow applications that receive and send messages.
 
-**Warning**: Wiki and all the docs refer to the 1.2.0.beta4. Sorry for the inconvenience. We will release the stable 1.2.0 version soon.
-
 ## How does it work
 
 Karafka provides a higher-level abstraction that allows you to focus on your business logic development, instead of focusing on implementing lower level abstraction layers. It provides developers with a set of tools that are dedicated for building multi-topic applications similar to how Rails applications are being built.