--- a/lib/karafka/helpers/inflector.rb
+++ b/lib/karafka/helpers/inflector.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Karafka
+  module Helpers
+    module Inflector
+      ENGINE = Dry::Inflector.new
+
+      def self.underscore(element)
+        ENGINE.underscore(element.to_s).sub('/', '_')
+      end
+    end
+  end
+end