--- a/lib/karafka/helpers/inflector.rb
+++ b/lib/karafka/helpers/inflector.rb
@@ -2,18 +2,23 @@
 
 module Karafka
   module Helpers
+    # Inflector provides inflection for the whole Karafka framework with additional inflection
+    # caching (due to the fact, that Dry::Inflector is slow)
     module Inflector
+      # What inflection engine do we want to use
       ENGINE = Dry::Inflector.new
 
       @map = Concurrent::Hash.new
 
-      class << self
-        def map(element)
-          @map[element] ||= ENGINE.underscore(element.to_s).tr('/', '_')
-        end
+      private_constant :ENGINE
 
-        def map!(attribute)
-          @map[attribute]
+      class << self
+        # @param string [String] string that we want to convert to our underscore format
+        # @return [String] inflected string
+        # @example
+        #   Karafka::Helpers::Inflector.map('Module/ControllerName') #=> 'module_controller_name'
+        def map(string)
+          @map[string] ||= ENGINE.underscore(string).tr('/', '_')
         end
       end
     end