--- a/lib/karafka/patches/ruby_kafka.rb
+++ b/lib/karafka/patches/ruby_kafka.rb
@@ -20,23 +20,26 @@ module Karafka
                       .select { |consumer| consumer.class.respond_to?(:after_fetch) }
 
           if Karafka::App.stopping?
-            publish_event(consumers,'before_stop')
+            publish_event(consumers, 'before_stop')
             Karafka::Persistence::Client.read.stop
           else
-            publish_event(consumers,'before_poll')
+            publish_event(consumers, 'before_poll')
             yield
-            publish_event(consumers,'after_poll')
+            publish_event(consumers, 'after_poll')
           end
         end
       end
 
       private
 
+      # Notifies consumers about particular events happening
+      # @param consumers [Array<Object>] all consumers that want to be notified about an event
+      # @param event_name [String] name of the event that happened
       def publish_event(consumers, event_name)
-        consumers.each { |consumer|
-          key = "consumers.#{Helpers::Inflector.map(consumer.class)}.#{event_name}"
+        consumers.each do |consumer|
+          key = "consumers.#{Helpers::Inflector.map(consumer.class.to_s)}.#{event_name}"
           Karafka::App.events.publish(key, context: consumer)
-        }
+        end
       end
     end
   end