--- a/lib/karafka/persistence/consumer.rb
+++ b/lib/karafka/persistence/consumer.rb
@@ -16,7 +16,7 @@ module Karafka
         def all
           # @note This does not need to be threadsafe (Hash) as it is always executed in a
           # current thread context
-          Thread.current[PERSISTENCE_SCOPE] ||= initial_state
+          Thread.current[PERSISTENCE_SCOPE] ||= Hash.new { |hash, key| hash[key] = {} }
         end
 
         # Used to build (if block given) and/or fetch a current consumer instance that will be
@@ -31,13 +31,6 @@ module Karafka
           else
             all[topic][partition] = topic.consumer.new
           end
-        end
-
-        private
-
-        # @return [Hash] initial, empty state for current thread's persistence scope
-        def initial_state
-          Hash.new { |hash, key| hash[key] = {} }
         end
       end
     end