--- a/lib/karafka/routing/consumer_mapper.rb
+++ b/lib/karafka/routing/consumer_mapper.rb
@@ -17,7 +17,7 @@ module Karafka
     #   module MyMapper
     #     def self.call(raw_consumer_group_name)
     #       [
-    #         Dry::Inflector.new.underscore(Karafka::App.config.client_id.to_s),
+    #         Karafka::Helpers::Inflector.map(Karafka::App.config.client_id.to_s),
     #         raw_consumer_group_name
     #       ].join('_').gsub('_', '.')
     #     end
@@ -26,7 +26,7 @@ module Karafka
       # @param raw_consumer_group_name [String, Symbol] string or symbolized consumer group name
       # @return [String] remapped final consumer group name
       def self.call(raw_consumer_group_name)
-        client_name = Dry::Inflector.new.underscore(Karafka::App.config.client_id.to_s)
+        client_name = Karafka::Helpers::Inflector.map(Karafka::App.config.client_id.to_s)
         "#{client_name}_#{raw_consumer_group_name}"
       end
     end