--- a/spec/lib/karafka/event_publisher_spec.rb
+++ b/spec/lib/karafka/event_publisher_spec.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-RSpec.describe Karafka::EventPublisher do
-  pending
-end