--- a/spec/lib/karafka/helpers/inflector_spec.rb
+++ b/spec/lib/karafka/helpers/inflector_spec.rb
@@ -4,13 +4,13 @@ RSpec.describe Karafka::Helpers::Inflector do
   describe '#map' do
     subject(:map) { described_class.map(string) }
 
-    context 'non namespaced controller name' do
+    context 'when we have string with a non namespaced controller name' do
       let(:string) { 'SomeFancyController' }
 
       it { is_expected.to eq 'some_fancy_controller' }
     end
 
-    context 'namespaced controller' do
+    context 'when we have string with a namespaced controller' do
       let(:string) { 'Namespaced::FancyController' }
 
       it { is_expected.to eq 'namespaced_fancy_controller' }