--- a/spec/lib/karafka/instrumentation/event_publisher_spec.rb
+++ b/spec/lib/karafka/instrumentation/event_publisher_spec.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+RSpec.describe Karafka::Instrumentation::EventPublisher do
+  pending
+end