--- a/spec/lib/karafka/instrumentation/event_publisher_spec.rb
+++ b/spec/lib/karafka/instrumentation/event_publisher_spec.rb
@@ -1,5 +1,10 @@
 # frozen_string_literal: true
 
 RSpec.describe Karafka::Instrumentation::EventPublisher do
-  pending
+  describe 'supported events' do
+    subject(:events) { described_class.events.keys }
+
+    it { expect(events).to include 'after_init' }
+    it { expect(events).to include 'before_fetch_loop' }
+  end
 end