… .ruby-version
… Gemfile.lock
… lib/karafka/event_publisher.rb