The end goal of this PR is to get rid of callbacks and introduce events with dry-events.

There are two types of callbacks:
1.) Home-baked for app-lifecycle events. The POC of rewriting them to dry-events is in this commit: https://github.com/karafka/karafka/commit/75737ccd6ba2d9119009df100120d42ac45baa33

2.) Activesupport based for consumer-lifecycle events `Kafka::Consumers::Callbacks`. 

We should keep backwards-compatibility with the previous behaviour. The main concern is dry-events uses payloads with `Hashes` whereas previously we relied on method signatures. 


This PR definitely needs more testing, but I prefer to get in sync early 😄

@mensfeld feedback plz 🙏 . Am I going into the right direction with this?