@madejejej 2018-03-28
Benchmark code:

```ruby
class CallbackedConsumer < ApplicationConsumer
  include Karafka::Consumers::Callbacks

  after_fetch { @x = 0 }

  before_stop { @x = 1 }

  before_poll { @x = 2 }

  after_poll { @x = 3 }

  def consume
  end
end

class NonCallbackedConsumerIncludeMixin < ApplicationConsumer
  include Karafka::Consumers::Callbacks

  def consume
  end
end

class NonCallbackedConsumer < ApplicationConsumer
  def consume
  end
end

# lib/tasks/bench.rake
# frozen_string_literal: true

task :bench do
  require 'benchmark/ips'
  consumer = CallbackedConsumer.new
  non_callbacked_consumer = NonCallbackedConsumer.new
  non_callbacked_consumer_include_mixin = NonCallbackedConsumerIncludeMixin.new

  App.consumer_groups.draw do
    consumer_group :batched_group do
      batch_fetching true

      topic :non_callbacked_consumer do
        consumer NonCallbackedConsumer
        batch_consuming true
      end

      topic :non_callbacked_consumer_include_mixin do
        consumer NonCallbackedConsumerIncludeMixin
        batch_consuming true
      end

      topic :callbacked_data do
        consumer CallbackedConsumer
        batch_consuming true
      end
    end
  end

  consumer.params_batch = []
  non_callbacked_consumer.params_batch = []
  non_callbacked_consumer_include_mixin.params_batch = []

  Benchmark.ips do |x|
    x.report("With Callbacks") { consumer.call }
    x.report("With Mixin Included") { non_callbacked_consumer_include_mixin.call }
    x.report("Without Mixin Included") { non_callbacked_consumer.call }
  end
end
```

```
> ruby -v
ruby 2.5.0p0 (2017-12-25 revision 61468) [x86_64-darwin17]
```

Karafka `rewrite-callbacks-to-dry-events` branch:
```
rake bench
Warming up --------------------------------------
      With Callbacks    14.145k i/100ms
 With Mixin Included    21.116k i/100ms
Without Mixin Included
                        25.578k i/100ms
Calculating -------------------------------------
      With Callbacks    157.752k (± 5.6%) i/s -    792.120k in   5.037354s
 With Mixin Included    220.684k (± 7.0%) i/s -      1.098M in   5.002665s
Without Mixin Included
                        294.512k (± 6.6%) i/s -      1.484M in   5.059950s
```

Karafka `master` branch:

```
Warming up --------------------------------------
      With Callbacks    10.265k i/100ms
 With Mixin Included	    20.725k i/100ms
Without Mixin Included
                        25.426k i/100ms
Calculating -------------------------------------
      With Callbacks    113.819k (± 7.6%) i/s -    574.840k in   5.081517s
 With Mixin Included    228.872k (± 5.1%) i/s -      1.161M in   5.084529s
Without Mixin Included
                        283.438k (± 7.5%) i/s -      1.424M in   5.053421s
```

MacBook Pro (15-inch, Late 2016)