@mensfeld 2018-04-09
@madejejej once you're back from your trips, could you: 

1. Retarget to 1.3-wip branch
2. Fix https://app.coditsu.io/karafka/builds/commit_builds/a757367b-41f6-4b6c-a3ac-3ae1c3c9c79a/validations

I guess I can apply some more remarks but those two things need to be done either way.