@madejejej 2018-04-09
@mensfeld Retargeted the branch.

The YARD violation is unremovable:

`Raised when an object is recognized but cannot be documented. This generally occurs when the Ruby syntax used to declare an object is too dynamic in nature.`

and it's because dry-events uses the module builder pattern.

I removed the other ones.