@mensfeld 2018-04-27
@madejejej remved the WIP for now. Also would be good to update with 1.3 when you will be working with it :)