@mensfeld 2018-05-04
@madejejej I've fixed the visibility problems. Feel free to benchmark.