@madejejej 2018-05-20
@mensfeld benchmarks

For the current version:

```
Warming up --------------------------------------
      With Callbacks     6.483k i/100ms
 With Mixin Included     6.220k i/100ms
Without Mixin Included
                        28.224k i/100ms
Calculating -------------------------------------
      With Callbacks     67.726k (±12.2%) i/s -    337.116k in   5.053245s
 With Mixin Included     62.532k (± 8.8%) i/s -    311.000k in   5.014042s
Without Mixin Included
                        303.951k (± 9.5%) i/s -      1.524M in   5.059467s
```

My last commit (10a11af)
 
```
Warming up --------------------------------------
      With Callbacks    16.476k i/100ms
 With Mixin Included    24.041k i/100ms
Without Mixin Included
                        30.501k i/100ms
Calculating -------------------------------------
      With Callbacks    178.292k (± 3.7%) i/s -    906.180k in   5.089617s
 With Mixin Included    260.591k (± 3.9%) i/s -      1.322M in   5.082329s
Without Mixin Included
                        345.659k (± 2.9%) i/s -      1.739M in   5.034322s
```

Looks like commit bdf7ff8 introduced slowed everything down:

```
Warming up --------------------------------------
      With Callbacks     7.318k i/100ms
 With Mixin Included     6.524k i/100ms
Without Mixin Included
                        31.409k i/100ms
Calculating -------------------------------------
      With Callbacks     78.455k (± 3.7%) i/s -    395.172k in   5.044549s
 With Mixin Included     66.720k (±10.5%) i/s -    332.724k in   5.049827s
Without Mixin Included
                        332.063k (± 6.9%) i/s -      1.665M in   5.039372s
```


Here's the benchmark source: https://gist.github.com/madejejej/a9956ce3955cf922ccff83e33edc9483