@mensfeld 2018-05-20
That is interesting. I will look into this shortly after I'm back from Japan.