@mensfeld 2018-05-20
@madejejej looked into this now. The results aren't that far now:

```
# Your last comment
With Callbacks    207.215k (± 6.5%) i/s -      1.033M in   5.008800s
With Mixin Included    314.171k (± 6.0%) i/s -      1.592M in   5.085102s
Without Mixin Included
                        419.958k (± 5.3%) i/s -      2.134M in   5.095775s

# My new local changes
With Callbacks    180.682k (± 2.5%) i/s -    904.233k in   5.007624s
With Mixin Included    234.302k (±12.4%) i/s -      1.148M in   5.006763s
Without Mixin Included
                        443.658k (± 7.1%) i/s -      2.231M in   5.060608s
```