@mensfeld 2018-05-20
@madejejej could you also rerun it on your hardware? I did some small optimizations and the results are much better.