@madejejej 2018-05-26
@mensfeld it's still around 30% slower than the previous version. What's funny is that it's a bit faster than the ActiveSupport version when the callbacks are running, but slower if we only include the mixin 😆 

```
Warming up --------------------------------------
      With Callbacks    13.752k i/100ms
 With Mixin Included    19.133k i/100ms
Without Mixin Included
                        33.506k i/100ms
Calculating -------------------------------------
      With Callbacks    148.305k (± 4.1%) i/s -    742.608k in   5.015653s
 With Mixin Included    200.209k (± 5.9%) i/s -      1.014M in   5.083639s
Without Mixin Included
                        353.321k (± 9.2%) i/s -      1.776M in   5.077140s
```