@mensfeld 2018-05-26
Hmmm,

Current head on this work:

```
Calculating -------------------------------------
      With Callbacks    148.305k (± 4.1%) i/s -    742.608k in   5.015653s
 With Mixin Included    200.209k (± 5.9%) i/s -      1.014M in   5.083639s
Without Mixin Included
                        353.321k (± 9.2%) i/s -      1.776M in   5.077140s
```

AS base (current master):

```
Warming up --------------------------------------
      With Callbacks     6.483k i/100ms
 With Mixin Included     6.220k i/100ms
Without Mixin Included
                        28.224k i/100ms
Calculating -------------------------------------
      With Callbacks     67.726k (±12.2%) i/s -    337.116k in   5.053245s
 With Mixin Included     62.532k (± 8.8%) i/s -    311.000k in   5.014042s
Without Mixin Included
                        303.951k (± 9.5%) i/s -      1.524M in   5.059467s
```

To me it seems, that it's 2-3 times faster than the current master for each case.